+++ lib/mixmod.rb
@@ -1,7 +1,6 @@
-require_relative './stat_tracker'
+require_relative "./stat_tracker"
 
 module MixMod
-
   def best_offense
     game_teams = @game_teams.data
     team_data = {}
@@ -272,77 +271,4 @@ module MixMod
     least_tackles = team_tackles.min_by { |k, v| v }
     least_tackles[0]
   end
-
-  def favorite_opponent(team_id)
-    goals_by_teams = Hash.new(0)
-    @teams.data.each do |team|
-      goals_by_teams[team[:team_id]] = {win: 0, total: 0, name: team[:teamname]} if
-      team[:team_id] != team_id
-    end
-    games_for_team_id = []
-    @games.data.each do |game|
-      if (game[:home_team_id] || game[:away_team_id]) == team_id
-        games_for_team_id << game[:game_id]
-      end
-    end
-    games_for_team_id.each do |expected|
-      @game_teams.data.each do |game|
-        if (game[:game_id] == expected) && (game[:team_id] != team_id)
-          if game[:result] == "WIN"
-            goals_by_teams[game[:team_id]][:win] += 1
-          end
-          goals_by_teams[game[:team_id]][:total] += 1
-        end
-      end
-    end
-    goals_by_teams.each do |team, hash|
-      hash[:percentage] = if hash[:total] == 0
-        1
-      else
-        (hash[:win].to_f / hash[:total].to_f)
-      end
-    end
-    goals_by_teams.delete(nil)
-    fav_opponent = goals_by_teams.min_by do |team, hash|
-      hash[:percentage]
-    end
-    fav_opponent[1][:name]
-  end
-
-  def rival(team_id)
-    goals_by_teams = Hash.new(0)
-    @teams.data.each do |team|
-      goals_by_teams[team[:team_id]] = {win: 0, total: 0, name: team[:teamname]} if
-      team[:team_id] != team_id
-    end
-    games_for_team_id = []
-    @games.data.each do |game|
-      if (game[:home_team_id] || game[:away_team_id]) == team_id
-        games_for_team_id << game[:game_id]
-      end
-    end
-    games_for_team_id.each do |expected|
-      @game_teams.data.each do |game|
-        if (game[:game_id] == expected) && (game[:team_id] != team_id)
-          if game[:result] == "WIN"
-            goals_by_teams[game[:team_id]][:win] += 1
-          end
-          goals_by_teams[game[:team_id]][:total] += 1
-        end
-      end
-    end
-    goals_by_teams.each do |team, hash|
-      hash[:percentage] = if hash[:total] == 0
-        0
-      else
-        hash[:percentage] = (hash[:win].to_f / hash[:total].to_f)
-      end
-    end
-    goals_by_teams.delete(nil)
-    rival_opponent = goals_by_teams.max_by do |team, hash|
-      hash[:percentage]
-    end
-    rival_opponent[1][:name]
-  end
-
 end
